+++ src/Events/Contracts/WithHttpRequest.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace AvtoDev\RoadRunnerLaravel\Events\Contracts;
+
+use Symfony\Component\HttpFoundation\Request;
+
+interface WithHttpRequest
+{
+    /**
+     * Get HTTP request instance.
+     *
+     * @return Request
+     */
+    public function httpRequest(): Request;
+}