+++ src/Listeners/BindRequestListener.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace AvtoDev\RoadRunnerLaravel\Listeners;
+
+use AvtoDev\RoadRunnerLaravel\Events\Contracts\WithApplication;
+use AvtoDev\RoadRunnerLaravel\Events\Contracts\WithHttpRequest;
+
+/**
+ * @link https://github.com/swooletw/laravel-swoole/blob/master/src/Server/Resetters/BindRequest.php
+ */
+class BindRequestListener implements ListenerInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function handle($event): void
+    {
+        if ($event instanceof WithApplication && $event instanceof WithHttpRequest) {
+            $event->application()->instance('request', $event->httpRequest());
+        }
+    }
+}